move non model utils to another module
