+++ ml_enabler/services/ml_model_service.py
@@ -1,7 +1,8 @@
 from flask import current_app
 from ml_enabler.models.ml_model import MLModel, MLModelVersion
 from ml_enabler.models.dtos.ml_model_dto import MLModelDTO, MLModelVersionDTO
-from ml_enabler.models.utils import NotFound, VersionNotFound, version_to_array
+from ml_enabler.models.utils import NotFound, VersionNotFound
+from ml_enabler.utils import version_to_array
 from sqlalchemy.orm.exc import NoResultFound
