+++ ml_enabler/services/prediction_service.py
@@ -1,10 +1,10 @@
 from flask import current_app
 from ml_enabler.models.ml_model import MLModel, MLModelVersion, Prediction, PredictionTile
 from ml_enabler.models.dtos.ml_model_dto import MLModelDTO, MLModelVersionDTO, PredictionDTO
-from ml_enabler.models.utils import NotFound, VersionNotFound,\
-    version_to_array, bbox_str_to_list, PredictionsNotFound, geojson_to_bbox,\
-    point_list_to_wkt, bbox_to_quadkeys, tuple_to_dict, polygon_to_wkt
-from sqlalchemy.orm.exc import NoResultFound
+from ml_enabler.models.utils import NotFound, VersionNotFound, \
+    PredictionsNotFound
+from ml_enabler.utils import bbox_to_quadkeys, tuple_to_dict, \
+    polygon_to_wkt, geojson_to_bbox, version_to_array, bbox_str_to_list
 from ml_enabler import db
